Move taslib source to lib/source.
